requirements.txt
Flask
torch
torchvision
Pillow
service_streamer
